move generic packages to thirdparty (see thirdparty/README.md)
