+++ blockservice/mock.go
@@ -6,7 +6,7 @@ import (
 	bitswap "github.com/jbenet/go-ipfs/exchange/bitswap"
 	tn "github.com/jbenet/go-ipfs/exchange/bitswap/testnet"
 	mockrouting "github.com/jbenet/go-ipfs/routing/mock"
-	delay "github.com/jbenet/go-ipfs/util/delay"
+	delay "github.com/jbenet/go-ipfs/thirdparty/delay"
 )
 
 // Mocks returns |n| connected mock Blockservices
