+++ p2p/net/swarm/swarm_listen.go
@@ -10,7 +10,7 @@ import (
 	context "github.com/jbenet/go-ipfs/Godeps/_workspace/src/code.google.com/p/go.net/context"
 	ma "github.com/jbenet/go-ipfs/Godeps/_workspace/src/github.com/jbenet/go-multiaddr"
 	ps "github.com/jbenet/go-ipfs/Godeps/_workspace/src/github.com/jbenet/go-peerstream"
-	multierr "github.com/jbenet/go-ipfs/util/multierr"
+	multierr "github.com/jbenet/go-ipfs/thirdparty/multierr"
 )
 
 // Open listeners for each network the swarm should listen on
