+++ p2p/protocol/relay/relay_test.go
@@ -8,7 +8,7 @@ import (
 	protocol "github.com/jbenet/go-ipfs/p2p/protocol"
 	relay "github.com/jbenet/go-ipfs/p2p/protocol/relay"
 	testutil "github.com/jbenet/go-ipfs/p2p/test/util"
-	eventlog "github.com/jbenet/go-ipfs/util/eventlog"
+	eventlog "github.com/jbenet/go-ipfs/thirdparty/eventlog"
 
 	context "github.com/jbenet/go-ipfs/Godeps/_workspace/src/code.google.com/p/go.net/context"
 )
