+++ repo/fsrepo/component/datastore_test.go
@@ -5,7 +5,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/jbenet/go-ipfs/repo/fsrepo/assert"
+	"github.com/jbenet/go-ipfs/thirdparty/assert"
 )
 
 // swap arg order
