+++ routing/mock/centralized_test.go
@@ -6,8 +6,8 @@ import (
 
 	context "github.com/jbenet/go-ipfs/Godeps/_workspace/src/code.google.com/p/go.net/context"
 	peer "github.com/jbenet/go-ipfs/p2p/peer"
+	delay "github.com/jbenet/go-ipfs/thirdparty/delay"
 	u "github.com/jbenet/go-ipfs/util"
-	delay "github.com/jbenet/go-ipfs/util/delay"
 	"github.com/jbenet/go-ipfs/util/testutil"
 )
