+++ util/datastore2/delayed.go
@@ -4,7 +4,7 @@ import (
 	ds "github.com/jbenet/go-ipfs/Godeps/_workspace/src/github.com/jbenet/go-datastore"
 	dsq "github.com/jbenet/go-ipfs/Godeps/_workspace/src/github.com/jbenet/go-datastore/query"
 
-	delay "github.com/jbenet/go-ipfs/util/delay"
+	delay "github.com/jbenet/go-ipfs/thirdparty/delay"
 )
 
 func WithDelay(ds ds.Datastore, delay delay.D) ds.Datastore {
